Version 1.6 Release Candidate 7
Adds config option for checkpoint system.
Makes data saving of player data async in most cases.

Only exception is when we create data for a new player or add new maps
to an existing player.

I will need to do quite a bit of redesign in order to make those async.